+++ app/src/main/kotlin/info/metadude/kotlin/library/roadsigns/demo/DemoActivity.kt
@@ -1,11 +1,11 @@
 package info.metadude.kotlin.library.roadsigns.demo
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.view.View
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import android.widget.Spinner
+import androidx.appcompat.app.AppCompatActivity
 import info.metadude.kotlin.library.roadsigns.RoadSign
 
 class DemoActivity : AppCompatActivity(), AdapterView.OnItemSelectedListener {